🏗️ Architect: Modularized skymap plotting into specialized package
Decomposed the `apts/plotting/skymap.py` God Object into the modular `apts/plotting/skymaps/` package.
Core dispatcher logic is now located in `apts/plotting/skymaps/base.py`.
The package maintains backward compatibility by re-exporting `plot_skymap` and `_generate_plot_skymap` in `apts/plotting/skymaps/__init__.py` and providing a legacy facade in the original `apts/plotting/skymap.py`.
Internal import chains and unit test mocks were updated to ensure consistency and correctness.
Also updated `tests/unit/test_iss_flybys_krakow.py` to match current orbital propagation data for April 2026.
