+++ apts/plotting/skymaps/__init__.py
@@ -3,6 +3,7 @@ from .skymap_polar import _generate_polar_skymap
 from .skymap_zoom import _generate_zoom_skymap
 from .skymap_texture import _generate_texture_skymap
 from .utils import setup_polar_ax, setup_zoom_ax
+from .base import plot_skymap, _generate_plot_skymap
 
 __all__ = [
     "resolve_target",
@@ -11,4 +12,6 @@ __all__ = [
     "_generate_texture_skymap",
     "setup_polar_ax",
     "setup_zoom_ax",
+    "plot_skymap",
+    "_generate_plot_skymap",
 ]
