+++ tests/plotting/test_equatorial_overlay.py
@@ -12,7 +12,7 @@ from apts.constants.plot import CoordinateSystem
 from apts.equipment import Equipment
 from apts.observations import Observation
 from apts.place import Place
-from apts.plotting.skymap import _generate_plot_skymap
+from apts.plotting.skymaps.base import _generate_plot_skymap
 
 
 @pytest.fixture
@@ -125,7 +125,7 @@ def test_equatorial_skymap_overlay_normal_azimuth(mock_obs_and_target):
 
     mock_observer.observe.side_effect = observe_side_effect
 
-    with patch("apts.plotting.skymap.pyplot") as mock_pyplot:
+    with patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot:
         mock_fig, mock_ax = MagicMock(), MagicMock()
         mock_pyplot.subplots.return_value = (mock_fig, mock_ax)
 
@@ -205,7 +205,7 @@ def test_equatorial_skymap_overlay_wrapping_azimuth(mock_obs_and_target):
 
     mock_observer.observe.side_effect = observe_side_effect
 
-    with patch("apts.plotting.skymap.pyplot") as mock_pyplot:
+    with patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot:
         mock_fig, mock_ax = MagicMock(), MagicMock()
         mock_pyplot.subplots.return_value = (mock_fig, mock_ax)
