+++ tests/plotting/test_plot_skymap.py
@@ -17,7 +17,7 @@ from apts.equipment import Equipment
 from apts.observations import Observation
 from apts.place import Place
 from apts.plotting import plot_skymap
-from apts.plotting.skymap import (
+from apts.plotting.skymaps.base import (
     _generate_plot_skymap,
 )
 from apts.plotting.skymap_objects import (
@@ -97,7 +97,7 @@ def mock_observation():
 def test_plot_skymap_renders_messier_objects(mock_observation):
     # Mock the necessary methods and data to avoid actual plotting
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch.object(Observation, "local_messier", MagicMock()) as mock_local_messier,
     ):
         # Mock the figure and axes objects
@@ -163,7 +163,7 @@ def test_plot_skymap_renders_messier_objects(mock_observation):
 def test_plot_ngc_object_with_no_size(mock_observation):
     # Mock the necessary methods and data to avoid actual plotting
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch(
             "apts.plotting.skymaps.skymap_zoom.get_brightness_color"
         ) as mock_get_brightness_color,
@@ -415,7 +415,7 @@ def test_plot_moon_on_skymap_renders_moon():
 
 def test_plot_planet_with_no_size_polar(mock_observation):
     # Mock the necessary methods and data to avoid actual plotting
-    with patch("apts.plotting.skymap.pyplot") as mock_pyplot:
+    with patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot:
         # Mock the figure and axes objects
         mock_fig = MagicMock()
         mock_ax = MagicMock()
@@ -776,7 +776,7 @@ def test_plot_messier_ellipse_angle_on_equatorial_zoom():
     )
 
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch(
             "apts.plotting.skymaps.skymap_zoom.calculate_parallactic_angle",
             return_value=parallactic_angle_val,
@@ -880,7 +880,7 @@ def test_plot_target_messier_ellipse_angle_on_horizontal_zoom():
     )
 
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch(
             "apts.plotting.skymaps.skymap_zoom.calculate_parallactic_angle",
             return_value=Angle(degrees=parallactic_angle_val),
@@ -1001,7 +1001,7 @@ def test_plot_non_target_messier_ellipse_angle_on_horizontal_zoom():
     )
 
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch(
             "apts.plotting.skymap_objects.calculate_parallactic_angle",
             return_value=Angle(degrees=parallactic_angle_val),
@@ -1121,7 +1121,7 @@ def test_plot_non_target_messier_ellipse_angle_on_equatorial_zoom():
     )
 
     with (
-        patch("apts.plotting.skymap.pyplot") as mock_pyplot,
+        patch("apts.plotting.skymaps.base.pyplot") as mock_pyplot,
         patch(
             "apts.plotting.skymap_objects.calculate_parallactic_angle",
             return_value=Angle(degrees=parallactic_angle_val),
